+++ src/CONTROLEUR/ChambreDAO.java
@@ -28,7 +28,7 @@ public class ChambreDAO extends DAO <Chambre> {
     
 
     @Override
-    public Chambre find(long id) {
+    public Chambre find(int id) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
